modularize handle_requirement
